refactor: move NotificationPopUp out of services and into views
